Refactos system metric executor
